YES! MIDDLEWARE 'MYAUTH' NOW IS WORKING CORRECTLY!
When user try to perform actions that are availeble only in authorized
mode, but he is not in it --
this middleware will redirect him to the page where
for him will propouse to make authorization or registration.

THE MAIN FEATURE in this process is button 'Back'.
How it works?
If user don't follow advice and click button back. It will
redirect him whenever he come from. Except no site urls and
currently this middleware url. In these cases he will redirected
to home page.